EPMLABSBRN-923 replace methods from StudyHistoryController to StudyHistoryControllerV2
